SDIT-873 Refactor Adjudication builders move DSL interface and rename top level DSL file
