+++ src/test/kotlin/uk/gov/justice/digital/hmpps/nomisprisonerapi/helper/builders/AdjudicationInvestigation.kt
@@ -1,10 +1,25 @@
 package uk.gov.justice.digital.hmpps.nomisprisonerapi.helper.builders
 
 import org.springframework.stereotype.Component
+import uk.gov.justice.digital.hmpps.nomisprisonerapi.jpa.AdjudicationEvidence
 import uk.gov.justice.digital.hmpps.nomisprisonerapi.jpa.AdjudicationIncidentParty
 import uk.gov.justice.digital.hmpps.nomisprisonerapi.jpa.AdjudicationInvestigation
 import uk.gov.justice.digital.hmpps.nomisprisonerapi.jpa.Staff
 import java.time.LocalDate
+
+@DslMarker
+annotation class AdjudicationInvestigationDslMarker
+
+@NomisDataDslMarker
+interface AdjudicationInvestigationDsl {
+  @AdjudicationEvidenceDslMarker
+  fun evidence(
+    detail: String = "Knife found",
+    type: String = "WEAP",
+    date: LocalDate = LocalDate.now(),
+    dsl: AdjudicationEvidenceDsl.() -> Unit = {},
+  ): AdjudicationEvidence
+}
 
 @Component
 class AdjudicationInvestigationBuilderFactory(
